Add persistence and integrate graph data into app state
- Move sample node data into a separate module
- Add persistence using localStorage for graph data
- Refactor graph data to be better incorporate into application state
