+++ .eslintrc.js
@@ -136,7 +136,7 @@ module.exports = {
         "no-continue": "error",
         "no-div-regex": "error",
         "no-duplicate-imports": "error",
-        "no-else-return": "error",
+        "no-else-return": "off",
         "no-empty-function": "error",
         "no-eq-null": "error",
         "no-eval": "error",
